Move useOpening to own file
