+++ packages/ui/src/app/pages/WorkingGroups/WorkingGroupsOpenings.tsx
@@ -5,7 +5,7 @@ import { ContentWithSidepanel, MainPanel, SidePanel } from '../../../common/comp
 import { Statistics } from '../../../common/components/statistics/Stats'
 import { Tabs } from '../../../common/components/Tabs'
 import { OpeningsList } from '../../../working-groups/components/OpeningsList'
-import { useOpenings } from '../../../working-groups/hooks/useWorkingGroups'
+import { useOpenings } from '../../../working-groups/hooks/useOpenings'
 import { AppPage } from '../../components/AppPage'
 
 type OpeningsTabs = 'OPENINGS' | 'UPCOMING'
