+++ packages/ui/src/working-groups/hooks/useWorkingGroups.ts
@@ -1,7 +1,5 @@
-import BN from 'bn.js'
-
 import { useGetWorkingGroupsQuery } from '../queries'
-import { asWorkingGroup, WorkingGroup, WorkingGroupOpening } from '../types'
+import { asWorkingGroup, WorkingGroup } from '../types'
 
 interface UseWorkingGroups {
   isLoading: boolean
@@ -16,38 +14,4 @@ export const useWorkingGroups = (): UseWorkingGroups => {
     isLoading: loading,
     groups: groups.map(asWorkingGroup),
   }
-}
-
-export const useOpenings = () => {
-  const openings: WorkingGroupOpening[] = [
-    {
-      id: '123',
-      title: 'Storage working group leader',
-      duration: [123, 'days'],
-      type: 'LEADER',
-      reward: { value: new BN(1000), interval: 3600 },
-      applicants: { current: 2, total: 10 },
-      hiring: { current: 0, total: 1 },
-    },
-    {
-      id: '221',
-      title: 'Storage working group worker',
-      duration: [12, 'days'],
-      type: 'REGULAR',
-      reward: { value: new BN(1000), interval: 3600 },
-      applicants: { current: 2, total: 10 },
-      hiring: { current: 0, total: 1 },
-    },
-    {
-      id: '2',
-      title: 'Membership group worker',
-      duration: [1, 'minutes'],
-      type: 'REGULAR',
-      reward: { value: new BN(1000), interval: 3600 },
-      applicants: { current: 0, total: 3 },
-      hiring: { current: 0, total: 1 },
-    },
-  ]
-
-  return openings
 }
